changed routes to match baseline, fixed code
